Automatically use context menus on iOS 13+
